+++ google-maps/src/implementation.ts
@@ -3,12 +3,15 @@ import { registerPlugin } from '@capacitor/core';
 
 import type {
   CameraConfig,
+  Circle,
   GoogleMapConfig,
   LatLng,
   LatLngBounds,
   MapPadding,
   MapType,
   Marker,
+  Polygon,
+  Polyline,
 } from './definitions';
 
 /**
@@ -55,6 +58,35 @@ export interface RemoveMarkersArgs {
 export interface AddMarkerArgs {
   id: string;
   marker: Marker;
+}
+
+export interface AddPolygonsArgs {
+  id: string;
+  polygons: Polygon[];
+}
+
+export interface RemovePolygonsArgs {
+  id: string;
+  polygonIds: string[];
+}
+
+export interface AddCirclesArgs {
+  id: string;
+  circles: Circle[];
+}
+
+export interface RemoveCirclesArgs {
+  id: string;
+  circleIds: string[];
+}
+export interface AddPolylinesArgs {
+  id: string;
+  polylines: Polyline[];
+}
+
+export interface RemovePolylinesArgs {
+  id: string;
+  polylineIds: string[];
 }
 
 export interface CameraArgs {
@@ -124,6 +156,12 @@ export interface CapacitorGoogleMapsPlugin extends Plugin {
   addMarkers(args: AddMarkersArgs): Promise<{ ids: string[] }>;
   removeMarker(args: RemoveMarkerArgs): Promise<void>;
   removeMarkers(args: RemoveMarkersArgs): Promise<void>;
+  addPolygons(args: AddPolygonsArgs): Promise<{ ids: string[] }>;
+  removePolygons(args: RemovePolygonsArgs): Promise<void>;
+  addCircles(args: AddCirclesArgs): Promise<{ ids: string[] }>;
+  removeCircles(args: RemoveCirclesArgs): Promise<void>;
+  addPolylines(args: AddPolylinesArgs): Promise<{ ids: string[] }>;
+  removePolylines(args: RemovePolylinesArgs): Promise<void>;
   enableClustering(args: EnableClusteringArgs): Promise<void>;
   disableClustering(args: { id: string }): Promise<void>;
   destroy(args: DestroyMapArgs): Promise<void>;