+++ google-maps/src/web.ts
@@ -29,6 +29,12 @@ import type {
   MapBoundsContainsArgs,
   EnableClusteringArgs,
   MapBoundsExtendArgs,
+  AddPolygonsArgs,
+  RemovePolygonsArgs,
+  AddCirclesArgs,
+  RemoveCirclesArgs,
+  AddPolylinesArgs,
+  RemovePolylinesArgs,
 } from './implementation';
 
 export class CapacitorGoogleMapsWeb
@@ -43,11 +49,23 @@ export class CapacitorGoogleMapsWeb
       markers: {
         [id: string]: google.maps.Marker;
       };
+      polygons: {
+        [id: string]: google.maps.Polygon;
+      };
+      circles: {
+        [id: string]: google.maps.Circle;
+      };
+      polylines: {
+        [id: string]: google.maps.Polyline;
+      };
       markerClusterer?: MarkerClusterer;
       trafficLayer?: google.maps.TrafficLayer;
     };
   } = {};
   private currMarkerId = 0;
+  private currPolygonId = 0;
+  private currCircleId = 0;
+  private currPolylineId = 0;
 
   private onClusterClickHandler: onClusterClickHandler = (
     _: google.maps.MapMouseEvent,
@@ -276,6 +294,93 @@ export class CapacitorGoogleMapsWeb
     delete this.maps[_args.id].markers[_args.markerId];
   }
 
+  async addPolygons(args: AddPolygonsArgs): Promise<{ ids: string[] }> {
+    const polygonIds: string[] = [];
+    const map = this.maps[args.id];
+
+    for (const polygonArgs of args.polygons) {
+      const polygon = new google.maps.Polygon(polygonArgs);
+      polygon.setMap(map.map);
+
+      const id = '' + this.currPolygonId;
+      this.maps[args.id].polygons[id] = polygon;
+      this.setPolygonListeners(args.id, id, polygon);
+
+      polygonIds.push(id);
+      this.currPolygonId++;
+    }
+
+    return { ids: polygonIds };
+  }
+
+  async removePolygons(args: RemovePolygonsArgs): Promise<void> {
+    const map = this.maps[args.id];
+
+    for (const id of args.polygonIds) {
+      map.polygons[id].setMap(null);
+      delete map.polygons[id];
+    }
+  }
+
+  async addCircles(args: AddCirclesArgs): Promise<{ ids: string[] }> {
+    const circleIds: string[] = [];
+    const map = this.maps[args.id];
+
+    for (const circleArgs of args.circles) {
+      const circle = new google.maps.Circle(circleArgs);
+      circle.setMap(map.map);
+
+      const id = '' + this.currCircleId;
+      this.maps[args.id].circles[id] = circle;
+      this.setCircleListeners(args.id, id, circle);
+
+      circleIds.push(id);
+      this.currCircleId++;
+    }
+
+    return { ids: circleIds };
+  }
+
+  async removeCircles(args: RemoveCirclesArgs): Promise<void> {
+    const map = this.maps[args.id];
+
+    for (const id of args.circleIds) {
+      map.circles[id].setMap(null);
+      delete map.circles[id];
+    }
+  }
+
+  async addPolylines(args: AddPolylinesArgs): Promise<{ ids: string[] }> {
+    const lineIds: string[] = [];
+    const map = this.maps[args.id];
+
+    for (const polylineArgs of args.polylines) {
+      const polyline = new google.maps.Polyline(polylineArgs);
+      polyline.set('tag', polylineArgs.tag);
+      polyline.setMap(map.map);
+
+      const id = '' + this.currPolylineId;
+      this.maps[args.id].polylines[id] = polyline;
+      this.setPolylineListeners(args.id, id, polyline);
+
+      lineIds.push(id);
+      this.currPolylineId++;
+    }
+
+    return {
+      ids: lineIds,
+    };
+  }
+
+  async removePolylines(args: RemovePolylinesArgs): Promise<void> {
+    const map = this.maps[args.id];
+
+    for (const id of args.polylineIds) {
+      map.polylines[id].setMap(null);
+      delete map.polylines[id];
+    }
+  }
+
   async enableClustering(_args: EnableClusteringArgs): Promise<void> {
     const markers: google.maps.Marker[] = [];
 
@@ -309,6 +414,9 @@ export class CapacitorGoogleMapsWeb
       map: new window.google.maps.Map(_args.element, { ..._args.config }),
       element: _args.element,
       markers: {},
+      polygons: {},
+      circles: {},
+      polylines: {},
     };
     this.setMapListeners(_args.id);
   }
@@ -357,6 +465,48 @@ export class CapacitorGoogleMapsWeb
       new google.maps.LatLng(_args.southwest.lat, _args.southwest.lng),
       new google.maps.LatLng(_args.northeast.lat, _args.northeast.lng),
     );
+  }
+
+  async setCircleListeners(
+    mapId: string,
+    circleId: string,
+    circle: google.maps.Circle,
+  ): Promise<void> {
+    circle.addListener('click', () => {
+      this.notifyListeners('onCircleClick', {
+        mapId: mapId,
+        circleId: circleId,
+        tag: circle.get('tag'),
+      });
+    });
+  }
+
+  async setPolygonListeners(
+    mapId: string,
+    polygonId: string,
+    polygon: google.maps.Polygon,
+  ): Promise<void> {
+    polygon.addListener('click', () => {
+      this.notifyListeners('onPolygonClick', {
+        mapId: mapId,
+        polygonId: polygonId,
+        tag: polygon.get('tag'),
+      });
+    });
+  }
+
+  async setPolylineListeners(
+    mapId: string,
+    polylineId: string,
+    polyline: google.maps.Polyline,
+  ): Promise<void> {
+    polyline.addListener('click', () => {
+      this.notifyListeners('onPolylineClick', {
+        mapId: mapId,
+        polylineId: polylineId,
+        tag: polyline.get('tag'),
+      });
+    });
   }
 
   async setMarkerListeners(